pyproject.toml
[project]
name = "spectral-take-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hyperliquid-python-sdk>=0.10.1",
    "ollama>=0.4.7",
    "pydantic>=2.10.6",
]
